+++ src/async_monitored_item.rs
@@ -17,14 +17,19 @@ use open62541_sys::{
 };
 use tokio::sync::mpsc;
 
-use crate::{ua, AsyncSubscription, CallbackOnce, CallbackStream, DataType as _, Error, Result};
+use crate::{
+    ua, AsyncSubscription, CallbackOnce, CallbackStream, DataType as _, Error, MonitoringFilter,
+    Result,
+};
 
 #[derive(Debug)]
 pub struct MonitoredItemBuilder {
     node_ids: Vec<ua::NodeId>,
+    attribute_id: Option<ua::AttributeId>,
     monitoring_mode: Option<ua::MonitoringMode>,
     #[allow(clippy::option_option)]
     sampling_interval: Option<Option<Duration>>,
+    filter: Option<Box<dyn MonitoringFilter>>,
     queue_size: Option<u32>,
     discard_oldest: Option<bool>,
 }
@@ -34,11 +39,24 @@ impl MonitoredItemBuilder {
     pub fn new(node_ids: impl IntoIterator<Item = ua::NodeId>) -> Self {
         Self {
             node_ids: node_ids.into_iter().collect(),
+            attribute_id: None,
             monitoring_mode: None,
             sampling_interval: None,
+            filter: None,
             queue_size: None,
             discard_oldest: None,
         }
+    }
+
+    /// Sets attribute ID.
+    ///
+    /// Default value is [`ua::AttributeId::VALUE`].
+    ///
+    /// See [`ua::MonitoredItemCreateRequest::with_attribute_id()`].
+    #[must_use]
+    pub fn attribute_id(mut self, attribute_id: ua::AttributeId) -> Self {
+        self.attribute_id = Some(attribute_id);
+        self
     }
 
     /// Sets monitoring mode.
@@ -63,7 +81,18 @@ impl MonitoredItemBuilder {
         self
     }
 
-    /// Set requested size of the monitored item queue.
+    /// Sets filter.
+    ///
+    /// Default value is no filter.
+    ///
+    /// See [`ua::MonitoringParameters::with_filter()`].
+    #[must_use]
+    pub fn filter(mut self, filter: impl MonitoringFilter) -> Self {
+        self.filter = Some(Box::new(filter));
+        self
+    }
+
+    /// Sets requested size of the monitored item queue.
     ///
     /// Default value is 1.
     ///
@@ -74,7 +103,7 @@ impl MonitoredItemBuilder {
         self
     }
 
-    /// Set discard policy.
+    /// Sets discard policy.
     ///
     /// Default value is `true`.
     ///
@@ -151,8 +180,10 @@ impl MonitoredItemBuilder {
     fn into_request(self, subscription_id: ua::SubscriptionId) -> ua::CreateMonitoredItemsRequest {
         let Self {
             node_ids,
+            attribute_id,
             monitoring_mode,
             sampling_interval,
+            filter,
             queue_size,
             discard_oldest,
         } = self;
@@ -162,11 +193,17 @@ impl MonitoredItemBuilder {
             .map(|node_id| {
                 let mut request = ua::MonitoredItemCreateRequest::default().with_node_id(&node_id);
 
+                if let Some(attribute_id) = attribute_id.as_ref() {
+                    request = request.with_attribute_id(attribute_id);
+                }
                 if let Some(monitoring_mode) = monitoring_mode.as_ref() {
                     request = request.with_monitoring_mode(monitoring_mode);
                 }
                 if let Some(&sampling_interval) = sampling_interval.as_ref() {
                     request = request.with_sampling_interval(sampling_interval);
+                }
+                if let Some(filter) = filter.as_ref() {
+                    request = request.with_filter(filter);
                 }
                 if let Some(&queue_size) = queue_size.as_ref() {
                     request = request.with_queue_size(queue_size);