+++ src/ua/data_types/monitored_item_create_request.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use open62541_sys::{UA_MonitoredItemCreateRequest_default, UA_NODEID_NUMERIC};
 
-use crate::{ua, DataType as _};
+use crate::{ua, DataType as _, MonitoringFilter};
 
 crate::data_type!(MonitoredItemCreateRequest);
 
@@ -60,6 +60,15 @@ impl MonitoredItemCreateRequest {
             } else {
                 -1.0
             };
+        self
+    }
+
+    /// Shortcut for setting filter.
+    #[must_use]
+    pub fn with_filter(mut self, filter: &impl MonitoringFilter) -> Self {
+        filter
+            .to_extension_object()
+            .move_into_raw(&mut self.0.requestedParameters.filter);
         self
     }