+++ CHANGELOG.md
@@ -7,6 +7,12 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add method `MonitoredItemBuilder::attribute_id()` to set a different attribute ID to monitor.
+- Add method `MonitoredItemBuilder::filter()` and related data types `ua::DataChangeFilter`,
+  `ua::EventFilter`, `ua::AggregateFilter` along with associated data types.
+
 ### Changed
 
 - Upgrade to open62541 version [1.4.9](https://github.com/open62541/open62541/releases/tag/v1.4.9).